+++ tests/InfluxDBClientTest.cs
@@ -667,5 +667,75 @@ namespace AdysTech.InfluxDB.Client.Net.Tests
             }
         }
 
+        [TestMethod, TestCategory("Post")]
+        public async Task TestPostPointsAsync_OlderthanRetention()
+        {
+            try
+            {
+                var client = new InfluxDBClient(influxUrl, dbUName, dbpwd);
+                var time = DateTime.Now;
+                var TestDate = time.ToShortDateString();
+                var TestTime = time.ToShortTimeString();
+                var points = new List<IInfluxDatapoint>();
+                InfluxRetentionPolicy retention = new InfluxRetentionPolicy() { Duration = TimeSpan.FromHours(1) };
+
+                for (var i = 0; i < 10; i++)
+                {
+                    await Task.Delay(1);
+                    var point = new InfluxDatapoint<long>();
+                    point.Retention = retention;
+                    point.UtcTimestamp = DateTime.UtcNow.AddDays(-i);
+                    point.MeasurementName = "RetentionTest";
+                    point.Precision = TimePrecision.Nanoseconds;
+                    point.Tags.Add("TestDate", TestDate);
+                    point.Tags.Add("TestTime", TestTime);
+                    point.Fields.Add("Val", i);
+                    points.Add(point);
+                }
+
+                var r = await client.PostPointsAsync(dbName, points);
+                Assert.IsTrue(r, "PostPointsAsync retunred false");
+
+                Assert.IsTrue(points.Count(p => p.Saved) == 1, "PostPointsAsync saved points older than retention policy");
+
+            }
+            catch (Exception e)
+            {
+                Assert.Fail($"Unexpected exception of type {e.GetType()} caught: {e.Message}");
+                return;
+            }
+        }
+
+        [TestMethod, TestCategory("Post")]
+        public async Task TestPostPointsAsync_DefaultTimePrecision()
+        {
+            try
+            {
+                var client = new InfluxDBClient(influxUrl, dbUName, dbpwd);
+                var time = DateTime.Now;
+                var TestDate = time.ToShortDateString();
+                var TestTime = time.ToShortTimeString();
+                var points = new List<IInfluxDatapoint>();
+                for (var i = 0; i < 10; i++)
+                {
+                    await Task.Delay(1);
+                    var point = new InfluxDatapoint<long>();
+                    point.MeasurementName = "PrecisionTest";
+                    point.Tags.Add("TestDate", TestDate);
+                    point.Tags.Add("TestTime", TestTime);
+                    point.Fields.Add("Val", i);
+                    points.Add(point);
+                }
+
+                var r = await client.PostPointsAsync(dbName, points);
+                Assert.IsTrue(r, "PostPointsAsync retunred false");
+            }
+            catch (Exception e)
+            {
+                Assert.Fail($"Unexpected exception of type {e.GetType()} caught: {e.Message}");
+                return;
+            }
+        }
+
     }
 }